Add GradingSchemeForm component
